requirements/local.txt
-r base.txt
django-debug-toolbar==1.2.2
django-extensions==1.4.9
